build for Linux.

+++ src/core/nodes.cpp
@@ -30,17 +30,17 @@ bool Nodes::add_zk_node(const zk_node& node) {
     auto it = m_zk_nodes.find(node.path());
     if (it == m_zk_nodes.end()) {
         m_zk_nodes[node.path()] = node;
-        for (int i = 1; i <= node.worker_cnt(); i++) {
+        for (size_t i = 1; i <= node.worker_cnt(); i++) {
             add_node(node.path(), node.type(), node.ip(), node.port(), i);
         }
     } else {
         auto old = it->second;
         if (old.SerializeAsString() != node.SerializeAsString()) {
             m_zk_nodes[node.path()] = node;
-            for (int i = 1; i < old.worker_cnt(); i++) {
+            for (size_t i = 1; i < old.worker_cnt(); i++) {
                 del_node(format_nodes_id(old.path(), old.ip(), old.port(), i));
             }
-            for (int i = 1; i <= node.worker_cnt(); i++) {
+            for (size_t i = 1; i <= node.worker_cnt(); i++) {
                 add_node(node.path(), node.type(), node.ip(), node.port(), i);
             }
             LOG_DEBUG("update zk node info done! path: %s", node.path().c_str());
@@ -58,7 +58,7 @@ bool Nodes::del_zk_node(const std::string& path) {
 
     /* delete nodes. */
     const zk_node& znode = it->second;
-    for (int i = 1; i <= znode.worker_cnt(); i++) {
+    for (size_t i = 1; i <= znode.worker_cnt(); i++) {
         del_node(format_nodes_id(znode.path(), znode.ip(), znode.port(), i));
     }
 
@@ -96,7 +96,7 @@ bool Nodes::add_node(const std::string& path, const std::string& node_type,
     node_t* node;
     std::vector<uint32_t> vnodes;
     VNODE2NODE_MAP& vnode2node = m_vnodes[node_type];
-    int old_vnode_cnt = vnode2node.size();
+    size_t old_vnode_cnt = vnode2node.size();
 
     vnodes = gen_vnodes(node_id);
     node = new node_t{node_id, node_type, ip, port, vnodes};
